requirements.txt
pydantic==2.11.9 
uvicorn==0.36.0
fastapi==0.117.1
fuzzywuzzy==0.18.0
websockets
python-socketio[client]
requests
Pillow
google-generativeai